Added afew new vehicles
Added Bus, Truck, and Motorbike.
Bus, max speed = 1, size = 3.
Truck, max speed = 2, size = 2.
Motorbike, max speed = 3, size = 1.
Motorbike will drive inbetween lanes if there is a vehicle in the way, i.e. if a bus is in front of it, it will continue its progress inbetween the lanes.

Expanded playing field to 4 * 2 lanes, and buffer is now 3 (to account for buses).