Added functionality to health bar and skill bar
Health bar will create 10 health icons at the start. When an enemy reaches the end, it will remove the health icons based on the enemy's damage.

Skill bar starts at 0, with a max capacity of 10. At the  start of every player's turn, a skill orb is gained. Every enemy kill also generates a skill orb.

At the moment, every unit have access to a single instance of GameStateManager to issue the damage values on crossing and orb values on death