Added skill orbs consumption and skill buttons interaction with orbs
Skill buttons will be disabled and greyed out when there is not enough skill orbs to use it.
Clicking on a skill will grey out the skill orbs on the skill orb bar, indicating what will be spent and what is left if the skill is used. At the same time, skill buttons will be disabled based on how many orbs are left if the selected skill is used.
Deselecting the skill, or cancelling the targeted skill will restore the greyed skill orbs to yellow.
Upon finishing player turn, all greyed orbs will be consumed and removed from skill orb bar. Skill buttons will be enabled/disabled accordingly on the next turn.

Note: At the moment, the code implemented for UI interaction with skill orbs and skill buttons could use some refactoring.