Edits to skill orb UI interaction
Selecting a skill will not lock in orbs deactivation and grey out the other skills yet, it will only update orb bar to show grey-ed if used. The deactivation is locked in only when a target is assigned and the skill is highlighted (queued for execution), then the skill bar will grey out the skills that now
 have insufficient orbs. Cancelling a highlighted skill will reactivate the grey orbs and refresh the skill bar.

The skill orb game objects in the skill orb bar are now all generated on startup, creating 10 active skill orbs and 10 inactive skill orbs, and these game objects will be disabled/enabled as required. Previous operation of instantiating and destroying the skill orbs has been removed.