Refactored skills related logic in UserControl, added ISkillManager interface
All skills now have their own Skill Manager class that implements ISkillManager. All logic related to the processing of the skill, targeting, skill UI management is all handled by the Skill Manager. UserControl will only handle skill buttons active/inactive and highlighting (for locked in/ not locked in), enabling/disabling buttons based on consumed skill orbs, and deselecting skill to cancel the execution of a locked in skill. On mouse left-click / right-click, if a skill is active, it will call the respective methods from ISkillManager to handle it. On player end turn, it will call each Skill Manager to execute their skill.

The Skill Manager themselves will keep track of whether it has been locked in or not.