Further polish, swap up to down
Switched enemy movement to spawn at the top and come downwards, with the ending point being the bottom sidewalk now.

Reworked vaulter. It now has the vault pole extended in front of it. When it first spawns, it will hold the vault vertically, in case there is a vehicle directly in front of its spawn. In this state, it is not vault ready, so it cannot vault. When its direct front is clear, it will bring the vault down and be vault ready onwards. In this state, the vault can be run over. Vaulter will now vault when a vehicle is 2 grids in front of it, and will vault 3 grids forward. Vaulting only usable once, then becomes soldier-like.

Several bug fixes, listing notable ones:
- lane changing will now correctly move units that are on top of the vehicle
- enemy units will reposition properly now, when spawning in and moving, with the Shield and Brute units arranged correctly to "visually" block vehicles